Organize materials processor package structure
