+++ src/materials_processor/__init__.py
@@ -1,0 +1,1 @@
+"""Materials Processor package."""
