+++ src/materials_processor/fixtures/__init__.py
@@ -1,0 +1,1 @@
+"""Materials Processor package."""
