+++ src/materials_processor/fixtures/example_output_nodes_dict.json
