+++ src/materials_processor/fixtures/example_traversed_nodes_dict.json
