+++ src/materials_processor/houdini/__init__.py
@@ -1,0 +1,1 @@
+"""Houdini integration for Materials Processor."""
