+++ src/materials_processor/usd/__init__.py
@@ -1,0 +1,1 @@
+"""USD integration for Materials Processor."""
